Fix a warning about toFOLD_utf8
In 5.30, toFOLD_utf8 will gain an extra parameter, denoting the end of
the input string (so that it doesn't try to read past it), and
to_utf8_fold (the older compatibility version we're using) will likely
go away entirely. As of 5.26, it warns and suggests using
toFOLD_utf8_safe (which has the parameter already and doesn't warn).

For 5.26 and up, use the _safe interface to silence the warning and
avoid breakage when 5.30 comes out. For older versions, continue using
to_utf8_fold to retain compatibility back to 5.8.

Fixes #7.